+++ test/unit/test_exception.py
@@ -8,6 +8,7 @@
 #
 ######################################################################
 
+from b2sdk.exception import FileOrBucketNotFound, ResourceNotFound
 from apiver_deps_exception import (
     AlreadyFailed,
     B2Error,
@@ -77,6 +78,13 @@ class TestInterpretError:
         self._check_one(FileNotPresent, 404, 'not_found', '', {})
         assert 'File not present: file.txt' == \
             str(interpret_b2_error(404, 'not_found', '', {}, {'fileName': 'file.txt'}))
+        assert 'File not present: 01010101' == \
+            str(interpret_b2_error(404, 'not_found', '', {}, {'fileId': '01010101'}))
+
+    def test_file_or_bucket_not_present(self):
+        self._check_one(ResourceNotFound, 404, None, None, {})
+        assert 'No such file, bucket, or endpoint: ' == \
+            str(interpret_b2_error(404, None, None, {}))
 
     def test_duplicate_bucket_name(self):
         self._check_one(DuplicateBucketName, 400, 'duplicate_bucket_name', '', {})