+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * Add custom exception for `403 transaction_cap_exceeded`
+* Add `get_file_info_by_id` and `get_file_info_by_name` to `Bucket`
+* `FileNotPresent` and `NonExistentBucket` now subclass new exceptions `FileOrBucketNotFound` and `ResourceNotFound`
 
 ### Changed
 * Fix missing import in the synchronization example